Various test fixes (mostly Windows-specific) (#15336)
