+++ tests/framework/console/controllers/PHPMessageControllerTest.php
@@ -120,6 +120,6 @@ class PHPMessageControllerTest extends BaseMessageControllerTest
         $content = file_get_contents($messageFilePath);
         $head = substr($content, 0, strpos($content, 'return '));
         $expected = "<?php\n/*file header*/\n/*doc block*/\n";
-        $this->assertSame($expected, $head);
+        $this->assertEqualsWithoutLE($expected, $head);
     }
 }
